feat(api server): server can creating n getting users, plus testing by swagger
